refactor(post): move submit post mutation logic into PostEditor folder
